LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-self-baseline-not-applied-if-sizing-cyclic-dependency-002-expected.txt

auto-sized rows - horizonal grid and verticalLR item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 min-content-sized rows - horizonal grid and verticalLR item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-content-sized rows - horizonal grid and verticalLR item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 fit-content-sized rows - horizonal grid and verticalLR item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 flex-sized rows - horizonal grid and verticalLR item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-flex-sized rows - horizonal grid and verticalLR item - column-axis baseline
baseline is not applied initially, but orthogonal items force repeating the track sizing and height is not indefinite in that phase.
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 auto-sized columns - horizontal grid item - row-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 min-content-sized columns - horizontal grid item - row-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-content-sized columns - horizontal grid item - row-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 fit-content-sized columns - horizontal grid item - row-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 flex-sized columns - horizontal grid item - row-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-flex-sized columns - horizontal grid item - row-axis baseline
baseline is not applied initially, but orthogonal items force repeating the track sizing and height is not indefinite in that phase.
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 auto-sized rows - verticalLR grid and horizontal item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 min-content-sized rows - verticalLR grid and horizontal item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-content-sized rows - verticalLR grid and horizontal item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 fit-content-sized rows - verticalLR grid and horizontal item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 flex-sized rows - verticalLR grid and horizontal item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-flex-sized rows - verticalLR grid and horizontal item - column-axis baseline
baseline is not applied initially, but orthogonal items force repeating the track sizing and height is not indefinite in that phase.
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 auto-sized rows - verticalRL grid and horizontal item - column-axis baseline
baseline alignment is not supported, because the RL items do not share an alignment context with the LR items
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 min-content-sized rows - verticalRL grid and horizontal item - column-axis baseline
baseline alignment is not supported, because the RL items do not share an alignment context with the LR items
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-content-sized rows - verticalRL grid and horizontal item - column-axis baseline
baseline alignment is not supported, because the RL items do not share an alignment context with the LR items
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 fit-content-sized rows - verticalRL grid and horizontal item - column-axis baseline
baseline alignment is not supported, because the RL items do not share an alignment context with the LR items
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 flex-sized rows - verticalRL grid and horizontal item - column-axis baseline
baseline alignment is not supported, because the RL items do not share an alignment context with the LR items
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-flex-sized rows - verticalRL grid and horizontal item - column-axis baseline
baseline alignment is not supported, because the RL items do not share an alignment context with the LR items
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É

PASS .inline-grid 1
PASS .inline-grid 2
PASS .inline-grid 3
PASS .inline-grid 4
PASS .inline-grid 5
PASS .inline-grid 6
PASS .inline-grid 7
PASS .inline-grid 8
PASS .inline-grid 9
PASS .inline-grid 10
PASS .inline-grid 11
PASS .inline-grid 12
PASS .inline-grid 13
PASS .inline-grid 14
PASS .inline-grid 15
PASS .inline-grid 16
PASS .inline-grid 17
PASS .inline-grid 18
PASS .inline-grid 19
PASS .inline-grid 20
PASS .inline-grid 21
PASS .inline-grid 22
PASS .inline-grid 23
FAIL .inline-grid 24 assert_equals:
<div class="inline-grid verticalRL alignItemsBaseline auto-columns max-flex-rows">
  <div class="firstRowFirstColumn bigFont paddingLeft" data-offset-x="366" data-offset-y="0" data-expected-width="75" data-expected-height="50">É</div>
  <div class="firstRowSecondColumn horizontalTB" data-offset-x="25" data-offset-y="50" data-expected-width="416" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 width25" data-offset-x="0" data-offset-y="0" data-expected-width="25" data-expected-height="150"></div>
</div>
offsetLeft expected 366 but got -10
